+++ src/global-cookies.js
@@ -3,7 +3,7 @@ const {
   COOKIE_HOSTS_USE_ROOT,
   AUTH_HOST,
   AUTH_PREFIX,
-} = require("./constants");
+} = require("./config/constants");
 const redirect = require("./redirect");
 
 const setGlobalCookies = (req, res, redirectUrl, cookies) => {
