+++ src/middlewares/request-logger.js
@@ -1,0 +1,8 @@
+module.exports = (req, res, next) => {
+  console.log(
+    `[${new Date().toISOString()}] - ${req.method} ${req.url} - ${
+      req.forwardedUri || "No forwarded URI"
+    }`
+  );
+  next();
+};
