+++ src/strategies.js
